Clean up some signed/unsigned comparisons

+++ src/libnet_advanced.c
@@ -81,7 +81,7 @@ libnet_adv_cull_header(libnet_t *l, libnet_ptag_t ptag, uint8_t **header,
 int
 libnet_adv_write_link(libnet_t *l, const uint8_t *packet, uint32_t packet_s)
 {
-    int c;
+    int32_t c;
 
     if (l->injection_type != LIBNET_LINK_ADV)
     {
@@ -89,10 +89,11 @@ libnet_adv_write_link(libnet_t *l, const uint8_t *packet, uint32_t packet_s)
                 "%s(): advanced link mode not enabled", __func__);
         return (-1);
     }
-    c = libnet_write_link(l, packet, packet_s);
+    // TODO: change to ssize_t?
+    c = (int32_t)libnet_write_link(l, packet, packet_s);
 
     /* do statistics */
-    if (c == packet_s)
+    if (c == (int32_t)packet_s)
     {
         l->stats.packets_sent++;
         l->stats.bytes_written += c;
@@ -115,7 +116,7 @@ libnet_adv_write_link(libnet_t *l, const uint8_t *packet, uint32_t packet_s)
 int
 libnet_adv_write_raw_ipv4(libnet_t *l, const uint8_t *packet, uint32_t packet_s)
 {
-    int c;
+    int32_t c;
 
     if (l->injection_type != LIBNET_RAW4_ADV)
     {
@@ -126,7 +127,7 @@ libnet_adv_write_raw_ipv4(libnet_t *l, const uint8_t *packet, uint32_t packet_s)
     c = libnet_write_raw_ipv4(l, packet, packet_s);
 
     /* do statistics */
-    if (c == packet_s)
+    if (c == (int32_t)packet_s)
     {
         l->stats.packets_sent++;
         l->stats.bytes_written += c;
